Fix storage implementation
