+++ Storage/src/linuxMain/kotlin/io/github/jan/supabase/storage/resumable/ResumableCacheUtil.kt
@@ -1,0 +1,6 @@
+package io.github.jan.supabase.storage.resumable
+
+import io.github.jan.supabase.annotations.SupabaseInternal
+
+@SupabaseInternal
+actual fun createDefaultResumableCache(): ResumableCache = MemoryResumableCache()
